feat(Quiz): create QuizManager component
